Fixed AGP 9.0 breaking changes
