+++ plugin/src/main/kotlin/com/github/gmazzo/buildconfig/BuildConfigTask.kt
@@ -3,8 +3,10 @@
 package com.github.gmazzo.buildconfig
 
 import com.github.gmazzo.buildconfig.generators.BuildConfigGeneratorSpec
+import com.github.gmazzo.buildconfig.internal.BuildConfigSourceSetInternal
 import java.io.File
 import org.gradle.api.DefaultTask
+import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.provider.SetProperty
@@ -13,6 +15,7 @@ import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
+import org.gradle.kotlin.dsl.newInstance
 import org.gradle.util.GradleVersion
 
 @CacheableTask
@@ -65,6 +68,30 @@ abstract class BuildConfigTask : DefaultTask() {
             )
         }
     }
+
+    internal fun bindTo(sourceSet: BuildConfigSourceSetInternal) = with(project) {
+        specs.add(provider { isolate(sourceSet) })
+        specs.addAll(provider { sourceSet.extraSpecs.map { isolate(it) } })
+    }
+
+    /**
+     * Helper method to create a copy of a BuildConfigClassSpec not linked to Gradle related extensions/DSL
+     * It makes it compatible with Configuration Cache
+     */
+    private fun Project.isolate(source: BuildConfigClassSpec) =
+        objects.newInstance<BuildConfigClassSpec>(source.name).apply spec@{
+            generator.value(source.generator).disallowChanges()
+            className.value(source.className).disallowChanges()
+            packageName.value(source.packageName).disallowChanges()
+            documentation.value(source.documentation).disallowChanges()
+            buildConfigFields.addAll(source.buildConfigFields.map { field ->
+                objects.newInstance<BuildConfigField>(field.name).apply field@{
+                    this@field.type.value(field.type).disallowChanges()
+                    this@field.value.value(field.value).disallowChanges()
+                    this@field.position.value(field.position).disallowChanges()
+                }
+            })
+        }
 
     private object HasFields : Spec<Task> {
         override fun isSatisfiedBy(task: Task): Boolean =
